+++ src/test/java/com/mailgrub/model/TaxTest.java
@@ -2,6 +2,8 @@ package com.mailgrub.model;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TaxTest {
@@ -11,10 +13,10 @@ public class TaxTest {
         Tax tax = new Tax();
         tax.setId(1);
         tax.setJurisdiction("Colorado");
-        tax.setTaxRate(7.25);
+        tax.setTaxRate(BigDecimal.valueOf(7.25));
 
         assertEquals(1, tax.getId());
         assertEquals("Colorado", tax.getJurisdiction());
-        assertEquals(7.25, tax.getTaxRate());
+        assertEquals(BigDecimal.valueOf(7.25), tax.getTaxRate());
     }
 }
